// src/types/game-state.d.ts
export interface NearbyStation {
  stationId: string;
  walkingTime: number;
}

export interface Station {
  id: string;
  name: string;
  coords: Coordinate;
  trackIds: string[];
  trackGroupId: string;
  buildType: BuildType;
  stNodeIds: string[];
  routeIds: string[];
  createdAt: number;
  nearbyStations: NearbyStation[];
}
// ...
export interface StNode {
  id: string;
  center: Coordinate;
  trackIds: string[];
  buildType: BuildType;
}

export interface SignalReference {
  signalId: string;
  areaCovered: 'all' | string;
}

export interface PathSegment {
  trackId: string;
  reversed: boolean;
  length: number;
  signals: SignalReference[];
}

export interface StCombo {
  startStNodeId: string;
  endStNodeId: string;
  path: PathSegment[];
  distance: number;
}

export interface StComboTiming {
  stNodeId: string;
  stNodeIndex: number;
  arrivalTime: number;
  departureTime: number;
}

export interface TrainSchedule {
  highDemand: number;
  mediumDemand: number;
  lowDemand: number;
}

export interface Route {
  id: string;
  name?: string;
  bullet?: string;
  color?: string;
  textColor?: string;
  shape?: RouteShape;
  carsPerTrain?: number;
  idealTrainCount?: number;
  trainType?: string;
  trainSchedule?: TrainSchedule;
  tempParentId?: string | null;
  stations?: Station[];
  stNodes?: StNode[];
  stCombos?: StCombo[];
  stComboTimings?: StComboTiming[];
}